Introduce the concept of incomplete volumes.
Incomplete volumes are individual members of multiple-disk setups. They
are now exposed as read-only devices (to make sure the user doesn't
damage the data which is part of a multi-device layout) with a special
partition hint. This is done to aid the user in data recovery, but these
partitions are normally not mountable by any conventional means.